Continued Working on Navigation Views #1 Added the conditional rendering for the back button on the navigation, to only display if you have moved forward in the view context. This was accomplished by wrapping a Frame inside of the NavigationView component, and binding it's 'CanGoBack' property to the back button's enabled property.
Also modified the logic which handles rendering the content to the NavView, to instead correctly render the view within the newly added Frame. (This allows us to have full metadata about the sequence of pages which have been viewed, in order to determine logic such as the conditional back button rendering). The new logic now programmatically resolves the proper uwp page from the '/Views' directory according to the 'Tag' property present on the NavigationViewItem entries.